requirements.txt
# Asset Shield V2 - Dependencies
# Python 3.11+ Required

# ============================================
# Core Data Processing
# ============================================
numpy>=1.24.0
pandas>=2.0.0
polars>=0.19.0

# ============================================
# Machine Learning / AI
# ============================================
scikit-learn>=1.3.0
torch>=2.0.0
google-generativeai>=0.3.0

# ============================================
# Financial Data Sources
# ============================================
# J-Quants API (Primary - Production)
requests>=2.31.0

# yfinance (Legacy - Prototype only)
yfinance>=0.2.28

# ============================================
# Web Scraping / Browser Automation
# ============================================
selenium>=4.15.0
webdriver-manager>=4.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# ============================================
# Database
# ============================================
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# ============================================
# Security / Cryptography
# ============================================
cryptography>=41.0.0

# ============================================
# API / Web Framework
# ============================================
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# ============================================
# Testing
# ============================================
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# ============================================
# Code Quality
# ============================================
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
isort>=5.12.0

# ============================================
# Utilities
# ============================================
python-dotenv>=1.0.0
pyyaml>=6.0.0
tqdm>=4.66.0
psutil>=5.9.0
